+++ 3D-fluid-simulator/GraphicEngine/shaders/CalculateNDC.glsl
@@ -4,7 +4,7 @@ uniform mat4 model;
 uniform mat4 view;
 uniform mat4 projection;
 uniform vec2 viewport;
-uniform uint shapeRadius;
+uniform float shapeRadius;
 
 vec4 CalculateNDC(in vec3 position, in vec4 offset){
 	vec4 outPosition = vec4(position * shapeRadius + offset.xyz, 1.0);
